+++ config/task.default.php
@@ -2,24 +2,23 @@
 /**
  * task配置文件
  * 例子：
- * 'process_a' => [
+ * 'demo' => [
  *      'daemon' => '* * * * * *',//秒 分 时 日 月 周
  *      'filename' => '/usr/local/php/bin/php', //执行程序
  *      'params' => [],//执行程序参数
  *      'single' => true,//如果进程在运行，则不执行，只保持一个进程
- *      'single_kill_previous' => false, //如果设置为true，则检查到上一个进程还在执行时，会停止上一个进程，并启动新的进程
  *      'standard_ouput' => '', //标准输出
  *      'error_output' => '', // 错误输出
  *  ]
  */
 
 return [
-    'process_a' => [
+    'demo' => [
         'daemon' => '*/3 * * * * *',
-        'filename' => '/usr/local/php-5.6.30/bin/php',
-        'params' => ['/www/tests/pcntl/examples/a.php'],
+        'filename' => 'echo',
+        'params' => ['"hello world!"'],
         'single' => true,
-        'standard_ouput' => '/www/tests/pcntl/examples/a.log',
-        'error_output' => '/www/tests/pcntl/examples/a.log',
+        'standard_ouput' => '/dev/null',
+        'error_output' => '/dev/null',
     ]
 ];
